contest model　追加
