+++ frontend/src/model/contest.ts
@@ -1,0 +1,6 @@
+export interface Contest {
+  category: string;
+  title: string;
+  startAt: string;
+  status: "upcoming" | "open" | "closed" | "archived";
+}
